feat: implement card reordering and update related schemas and services refactor: adjust list and card repository queries for better performance fix: update authorization handling in routes and add missing auth service

+++ src/entities/card.entity.ts
@@ -18,7 +18,7 @@ export class Card extends DateTimeEntity {
     @Column({ type: 'text', nullable: true })
     description: string
 
-    @Column({ type: 'int', default: 0 })
+    @Column({ type: 'float', default: 0 })
     position: number
 
     @Column({ type: 'varchar', length: 255, nullable: true })
